drivers: mfd: axonfabric: Add debugfs, prepere irq support, lock-direction
Add debugfs support to enable reading all configuration registers from
userspace.

Prepare IRQ support by adding regsiter definitions for the IO bank
IRQ registers.

Add lock-direction support. This is useful when performing loopback
testing of the IO between the SOC and fabric. When lock-direction
is configured in the DT, the gpio calls will not change the direction
of the IO blocks. If configured as inputs, then the ioblocks will
only be inputs (outputs to the SOC). Then the ODR can be used to
set a high or low level on the IO block, and that can be read
by the corrseponding SOC gpio pad.  Care must be taken to not cause
a contention situation (where the SOC drives the IO at the same time
as the fabric).