drivers:mfd:axonfabric: Turn off regmap cache and really add reset
Disabling regmap cache to avoid an intermittent error with fabric initialization.

The error symptom is that the device probe fails to read the magic word. The device
returns all 0x00 instead of the magic word.

Also, really perform a device reset. Previous patch didn't store the reset gpio
descriptor, so this commit fixes that issue.